refactor(marketplace): land it as a workspace crate, start the contract split
Moves the assembled plugin to `crates/renzora_marketplace` and finishes it.
The four engine integrations it could not reach as a native plugin -- audio,
theme, shader and grid previews -- all resolve now, and the crate compiles,
lints and stages clean.

Two of the four became contracts on the way, because they are worth having
regardless of where this crate lives:

* `renzora::grid` -- `InfiniteGrid` and `InfiniteGridSettings`. Pure data, a
  marker and a struct of colours and floats, so the contract crate gains no
  dependency. `renzora_grid` keeps the pipeline and re-exports both. Any plugin
  can now put a ground plane under its own preview.
* `renzora::audio` -- `AudioLink` and the handle types. This one cost almost
  nothing: `link.rs` had zero references to the rest of `renzora_audio` (it was
  already a self-contained boundary), and the play/stop vocabulary it speaks was
  always in `renzora_plugin::audio`, which the contract crate already depended on
  for the HTTP `net` types. Kira, the mixer, emitters and the timeline all stay
  put.

Both sit behind cargo features, like `text_mesh`, so a lean export compiles only
what it uses. Theme and shader stay where they are for now: moving them would
mean moving implementations, which is the line the contract crate should not
cross.

Become-a-Creator is folded into Publish. They were two panels, and the split
never made sense -- the uploader's only useful message to a non-creator was to go
and open the other one. Both stages are built up front and swapped by
`bind_display` on a tracked read of creator status, so a background status poll
switches them with no rebuild and no lost form state. The gate is deliberately
looser than "fully onboarded": someone who chose "Skip for now" on the payout
step can still publish, just nothing paid.

Also removes what the deleted panels left behind -- per-area accent hues, the
post action chip, relative timestamps, local moderation notes, role icons,
presence avatars, cover-fill images and two API helpers -- and fixes two
pre-existing clippy failures in `renzora_export` that the CI gate was already
tripping on, one of which was a `format!` whose escaped braces would have been
emitted literally once it stopped interpolating.

Docs: the contract-vs-implementation rule is written down in
`extending/native-plugins.md`, with the test for whether a candidate belongs.

+++ crates/renzora_export/src/native.rs
@@ -629,7 +629,7 @@ fn rebuild_right_pane(world: &mut World) {
         .query_filtered::<&bevy::window::Window, With<bevy::window::PrimaryWindow>>()
         .iter(world)
         .next()
-        .map(|w| w.resolution.height() as f32 / w.resolution.scale_factor())
+        .map(|w| w.resolution.height() / w.resolution.scale_factor())
         .unwrap_or(0.0);
     let tab_max = tab_content_max(window_height);
     let mut queue = CommandQueue::default();
